Migrate enterprise commands to upbound
Migrates all enterprise commands to instead fall under the upbound
subcommands. Minor terminology updates are made in flag descriptions as
well.
